change: 世間とのずれをhooksに
