+++ src/pages/solo/fin.tsx
@@ -9,26 +9,7 @@ import { useRecoilState } from 'recoil';
 import { motion } from 'framer-motion';
 import { hoverTapLink } from '@/animations/variants';
 import { ItemData } from '@/types/Game';
-
-const calcItemDiffPercentage = (
-  itemPrice: number[],
-  ansPrice: number[]
-) => {
-  // itemと入力値の差の割合を出力する
-  const diffPercent =
-    itemPrice.reduce(
-      (prev, curr, index) =>
-        prev + Math.abs(ansPrice[index]) / Math.abs(curr),
-      0
-    ) * 100;
-
-  const ret = Math.round(diffPercent);
-  return ret;
-};
-
-const getItemToPrice = (item: ItemData[]) => {
-  return item.map((itemData) => itemData.answer);
-};
+import { useItemDiffPercentage } from '@/hooks/useItemDiffPerceentage';
 
 const Fin = () => {
   const [item, setItem] = useRecoilState(itemData);
@@ -37,10 +18,7 @@ const Fin = () => {
     keyPadNumArrState
   );
 
-  const itemDiffPercentage = calcItemDiffPercentage(
-    getItemToPrice(item),
-    keyPadNumArr
-  );
+  const itemDiffPercentage = useItemDiffPercentage();
 
   return (
     <div style={styles.container}>
@@ -52,14 +30,13 @@ const Fin = () => {
         <div style={styles.totalAns}>
           <h1>
             あなたの合計差額:
-            {keyPadNumArr.reduce(
-              (acc, num) => acc + Math.abs(num),
-              0
-            )}
+            {keyPadNumArr
+              .reduce((acc, num) => acc + Math.abs(num), 0)
+              .toLocaleString()}
           </h1>
           <h1>
             あなたの世間とのずれ:
-            {itemDiffPercentage}%
+            {itemDiffPercentage.toLocaleString()}%
           </h1>
         </div>
         {/* 上側 */}
@@ -89,7 +66,7 @@ const Fin = () => {
                   fontSize: 16,
                 }}
               >
-                金額：{item.answer}円
+                金額：{item.answer.toLocaleString()}円
               </p>
               <h1
                 style={{
@@ -101,7 +78,7 @@ const Fin = () => {
               >
                 差額金額：
                 <br />
-                {keyPadNumArr[index]}円
+                {keyPadNumArr[index].toLocaleString()}円
               </h1>
               <motion.a
                 href={item.affiliatelink}
